fully functional setup

+++ computer.cc
@@ -5,7 +5,7 @@
 using namespace std;
 
 
-Computer_Lv1::Computer_Lv1(int who, Game *g, int level, int score, King *king): Computer{who, g, level, score, king} {}
+Computer_Lv1::Computer_Lv1(int who, Game *g, int level, float score, King *king): Computer{who, g, level, score, king} {}
 
 Computer_Lv1::~Computer_Lv1() {}
 
@@ -45,7 +45,7 @@ vector<string> Computer_Lv1::smartMove() {
     return pureRandom(getGame());
 }
 
-Computer_Lv2::Computer_Lv2(int who, Game *g, int level, int score, King *king): Computer{who, g, level, score, king} {}
+Computer_Lv2::Computer_Lv2(int who, Game *g, int level, float score, King *king): Computer{who, g, level, score, king} {}
 Computer_Lv2::~Computer_Lv2() {}
 
 vector<string> Computer_Lv2::smartMove() {
@@ -90,13 +90,13 @@ vector<string> Computer_Lv2::smartMove() {
     }
 }
 
-Computer_Lv3::Computer_Lv3(int who, Game *g, int level, int score, King *king): Computer{who, g, level, score, king} {}
+Computer_Lv3::Computer_Lv3(int who, Game *g, int level, float score, King *king): Computer{who, g, level, score, king} {}
 
 Computer_Lv3::~Computer_Lv3() {}
 
 vector<string> Computer_Lv3::smartMove() {}
 
-Computer_Lv4::Computer_Lv4(int who, Game *g, int level, int score, King *king): Computer{who, g, level, score, king} {}
+Computer_Lv4::Computer_Lv4(int who, Game *g, int level, float score, King *king): Computer{who, g, level, score, king} {}
 
 Computer_Lv4::~Computer_Lv4() {}
