+++ computer.h
@@ -7,21 +7,21 @@
 
 class Computer_Lv1 : public Computer {
     public:
-    Computer_Lv1(int, Game *, int, int = 0, King * = nullptr);
+    Computer_Lv1(int, Game *, int, float = 0, King * = nullptr);
     ~Computer_Lv1();
     virtual std::vector<std::string> smartMove() override final;
 };
 
 class Computer_Lv2 : public Computer {
     public:
-    Computer_Lv2(int, Game *, int, int = 0, King * = nullptr);
+    Computer_Lv2(int, Game *, int, float = 0, King * = nullptr);
     ~Computer_Lv2();
     virtual std::vector<std::string> smartMove() override final;
 };
 
 class Computer_Lv3 : public Computer {
     public:
-    Computer_Lv3(int, Game *, int, int = 0, King * = nullptr);
+    Computer_Lv3(int, Game *, int, float = 0, King * = nullptr);
     ~Computer_Lv3();
     virtual std::vector<std::string> smartMove() override final;
 };
@@ -29,7 +29,7 @@ class Computer_Lv3 : public Computer {
 
 class Computer_Lv4 : public Computer {
     public:
-    Computer_Lv4(int, Game *, int, int = 0, King * = nullptr);
+    Computer_Lv4(int, Game *, int, float = 0, King * = nullptr);
     ~Computer_Lv4();
     virtual std::vector<std::string> smartMove() override final;
 };
